# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Finance Tracker API",
      "description": "RESTful API for tracking personal finances",
      "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
      "http"
    ],
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/budgets": {
        "get": {
          "summary": "Get all budgets",
          "tags": ["Budgets"],
          "security": [{"Bearer": []}],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "query",
              "name": "active_only",
              "type": "boolean",
              "description": "Filter active budgets only"
            },
            {
              "in": "query",
              "name": "category_id",
              "type": "integer",
              "description": "Filter by category ID"
            }
          ],
          "responses": {
            "200": {
              "description": "List of budgets"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "post": {
          "summary": "Create a new budget",
          "tags": ["Budgets"],
          "security": [{"Bearer": []}],
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "category_id": {
                    "type": "integer"
                  }
                },
                "required": ["amount", "start_date", "end_date", "category_id"]
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Budget created successfully"
            },
            "400": {
              "description": "Invalid input"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Category not found"
            }
          }
        }
      },
      "/budgets/{id}": {
        "get": {
          "summary": "Get a specific budget",
          "tags": ["Budgets"],
          "security": [{"Bearer": []}],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "integer",
              "description": "Budget ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Budget details"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Budget not found"
            }
          }
        },
        "put": {
          "summary": "Update a budget",
          "tags": ["Budgets"],
          "security": [{"Bearer": []}],
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "integer",
              "description": "Budget ID"
            },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "category_id": {
                    "type": "integer"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Budget updated successfully"
            },
            "400": {
              "description": "Invalid input"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Budget or category not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a budget",
          "tags": ["Budgets"],
          "security": [{"Bearer": []}],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "integer",
              "description": "Budget ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Budget deleted successfully"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Budget not found"
            }
          }
        }
      },
...
    }
  }